+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # DocuSign Java Client Changelog
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
-## [v3.10.0-RC1] - eSignature API v2.1-20.3.01 - 2020-10-15
+## [v3.10.1] - eSignature API v2.1-20.3.01 - 2020-10-29
 ### Changed
 - Added support for version v2.1-20.3.01 of the DocuSign eSignature API.
 - Updated the SDK release version.
 - Updated ApiClient to use an empty JSON object if the body is null. (DCM-4359)
+### Fixed
+- AccountBrands::updateBrandResourcesByContentType is missing brand file option. (DCM-3868)
+
+## [v3.10.0-RC1] - eSignature API v2.1-20.3.01 - 2020-10-15
+### Changed
+- Added support for version v2.1-20.3.01 of the DocuSign eSignature API.
+- Updated the SDK release version.
 
 ## [v3.9.0] - eSignature API v2.1-20.3.00 - 2020-09-24
 ### Changed